// frontend/src/types/api.ts
export interface components {
    schemas: {
// ...
        InvoiceResponse: {
            /**
             * Format: uuid
             * @description Invoice ID
             * @example 550e8400-e29b-41d4-a716-446655440000
             */
            id?: string;
            /**
             * Format: uuid
             * @description Customer ID
             * @example 550e8400-e29b-41d4-a716-446655440000
             */
            customerId?: string;
            /**
             * @description Invoice status
             * @example DRAFT
             */
            status?: string;
            /**
             * @description Total amount
             * @example 1000
             */
            totalAmount?: number;
            /**
             * @description Current balance
             * @example 1000
             */
            balance?: number;
            /** @description Line items */
            lineItems?: components["schemas"]["LineItemResponse"][];
            /**
             * Format: date-time
             * @description Creation timestamp
             */
            createdAt?: string;
            /**
             * Format: date-time
             * @description Last update timestamp
             */
            updatedAt?: string;
        };
// ...
        PageInvoiceResponse: {
            /** Format: int64 */
            totalElements?: number;
            /** Format: int32 */
            totalPages?: number;
            pageable?: components["schemas"]["PageableObject"];
            /** Format: int32 */
            numberOfElements?: number;
            first?: boolean;
            last?: boolean;
            /** Format: int32 */
            size?: number;
            content?: components["schemas"]["InvoiceResponse"][];
            /** Format: int32 */
            number?: number;
            sort?: components["schemas"]["SortObject"];
            empty?: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}